backend/requirements.txt
Flask
flask-cors
python-dotenv
google-genai
google-api-core
tenacity
PyMuPDF
Pillow
pytesseract
gunicorn # Used by Render to run the Flask app